Move EIP712 logic to its own module
