+++ libs/ledgerjs/packages/hw-app-eth/src/modules/EIP712/EIP712.types.ts
